phase2_langchain/requirements.txt
# ============================================
# Phase 2: LangChain - 依赖包
# ============================================
# 安装: pip install -r requirements.txt

# ============================================
# LangChain 核心
# ============================================
langchain==0.1.5
langchain-core==0.1.13

# ============================================
# LangChain 社区集成
# ============================================
langchain-community==0.0.19

# ============================================
# OpenAI 集成
# ============================================
langchain-openai==0.0.5

# ============================================
# Anthropic 集成
# ============================================
langchain-anthropic==0.1.8

# ============================================
# 文档处理
# ============================================
pypdf==3.17.4
python-docx==1.1.0
beautifulsoup4==4.12.3
unstructured==0.12.5

# ============================================
# 向量数据库
# ============================================
chromadb==0.4.22
faiss-cpu==1.7.4

# ============================================
# 嵌入模型
# ============================================
sentence-transformers==2.3.1

# ============================================
# 工具库
# ============================================
tiktoken==0.5.2
python-dotenv==1.0.0
